072724 (for main blog) added character encryption effect for blog tittles! + added data processing sound effect too with jukebox + added sound effect for back buttons (tried for cards too but not working real well rn). + figured out the styling for the blogs without using is:global and it effecting the rest of the site - I just added the Blog layout to each entry and did the CSS there and it worked! no more nbsp! + added tag slugs and sites + added tag filters on main writing page. all static.

+++ src/utils.ts
@@ -2,11 +2,11 @@
 function formatDate(date: Date): string {
     const options: Intl.DateTimeFormatOptions = {
       year: 'numeric',
-      month: 'long',
+      month: 'short',
       day: 'numeric',
     };
   
     return new Date(date).toLocaleDateString(undefined, options);
   }
-  
+
 export { formatDate };
